requirements.txt
amqp==2.4.2
beautifulsoup4==4.7.1
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
crontab==0.22.5
Flask==1.0.2
future==0.17.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.5.0
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
pytz==2019.1
requests==2.21.0
soupsieve==1.9.1
urllib3==1.24.2
vine==1.3.0
Werkzeug==0.15.2
